feat(unidentified-key-types): Added index.d for typescript interfaces… (#9)
* feat(unidentified-key-types): Added index.d for typescript interfaces and refactored code

* feat(unidentified-key-types): Updated project files
